requirements.txt
mysqlclient==2.1.1
python-dotenv==0.21.0
web3==5.30.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1

coverage==6.4.4
flake8==5.0.4
pytest-cov==3.0.0
pytest==7.1.3